+++ shared/@common/ui/Button/ButtonComponent.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Table from '@/shared/@common/ui/Table/Table';
+import Table from '@/shared/@common/ui/Table/TableBody';
 import Button from '@/shared/@common/ui/Button/Button';
 
 const ButtonComponent = () => {
